+++ tests/testthat/test-mng01.R
@@ -60,8 +60,10 @@ test_that("mng01 works with combination of x variables", {
     proc_data,
     dataset = "adlb",
     x_var = c("AVISIT", "AVISITN"),
-    show_v_grid = TRUE,
-    show_h_grid = TRUE
+    ggtheme = gg_theme_chevron(
+      grid_y = TRUE,
+      grid_x = TRUE
+    )
   )
   expect_list(res, len = 3, types = "ggplot")
 })